+++ examples/basic.rs
@@ -12,5 +12,5 @@ fn main() {
 // `Audio` is an alias for `AudioChannel<MainTrack>`, which is the default channel added by the audio plugin
 // See the `custom_channel` example to add your own audio channels
 fn play_loop(asset_server: Res<AssetServer>, audio: Res<Audio>) {
-    audio.play_looped(asset_server.load("sounds/loop.ogg"));
+    audio.play(asset_server.load("sounds/loop.ogg")).looped();
 }